refactor(action): extract checks file
